EDA backend   Django/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.15
dj-database-url>=1.0
psycopg2-binary>=2.9
